feat(main): add capability to wait for model synchronization
This commit adds the resource that waits for model to get to active or
error state. This can controlled by a boolean variable `wait_for_model`
whose default value is kept as false to run the plan even on machines
with out `juju-wait` as this resource requires the dependency to be
installed first.