feat!: refactor the plan according to subcluster and controller models
This commit changes the plan to more closely resemble what we deploy for
customers in terms of the juju topology. The new plan has two
submodules, one for subcluster and the other for controller. The two
models are connected to each other using cross models. The juju offers
are being exposed by each terraform module to connect the two
corresponding models.